fix(coding-agent): keep service tier policy out of startup imports
