+++ packages/coding-agent/src/config/service-tier.ts
@@ -1,12 +1,4 @@
-import {
-	serviceTierFamily,
-	type Model,
-	type ServiceTier,
-	type ServiceTierByFamily,
-	type ServiceTierFamily,
-} from "@oh-my-pi/pi-ai";
-import type { ThinkingLevel } from "@oh-my-pi/pi-agent-core";
-import { toReasoningEffort } from "../thinking";
+import type { ServiceTier, ServiceTierByFamily, ServiceTierFamily } from "@oh-my-pi/pi-ai/types";
 import type { SubmenuOption } from "./settings-schema";
 
 /**
@@ -195,39 +187,4 @@ export function validateServiceTierOverrides(value: unknown): Record<string, str
 		throw new Error(`Invalid tier.modelOverrides entries — ${invalid.join("; ")}`);
 	}
 	return validated;
-}
-
-export type ModelServiceTierOverrideResolution = { matched: false } | { matched: true; tier: ServiceTier | undefined };
-
-const NO_MODEL_SERVICE_TIER_OVERRIDE = Object.freeze({ matched: false as const });
-
-/**
- * Try provider/model:effort before provider/model, using the concrete model ID
- * verbatim. Off, inherit and undefined add no effort suffix. Unsupported family
- * values are inert. A matched "none" is authoritative off; no match leaves
- * the caller's baseline unchanged.
- */
-export function resolveModelServiceTierOverride(
-	overrides: Readonly<Record<string, string>>,
-	model: Model,
-	thinkingLevel: ThinkingLevel | undefined,
-): ModelServiceTierOverrideResolution {
-	const family = serviceTierFamily(model);
-	if (!family) return NO_MODEL_SERVICE_TIER_OVERRIDE;
-	const baseKey = `${model.provider}/${model.id}`;
-	const effort = toReasoningEffort(thinkingLevel);
-	if (effort) {
-		const effortKey = `${baseKey}:${effort}`;
-		if (Object.hasOwn(overrides, effortKey)) {
-			const tier = overrides[effortKey];
-			if (tier === "none") return { matched: true, tier: undefined };
-			if (isServiceTierForFamily(family, tier)) return { matched: true, tier };
-		}
-	}
-	if (Object.hasOwn(overrides, baseKey)) {
-		const tier = overrides[baseKey];
-		if (tier === "none") return { matched: true, tier: undefined };
-		if (isServiceTierForFamily(family, tier)) return { matched: true, tier };
-	}
-	return NO_MODEL_SERVICE_TIER_OVERRIDE;
 }
